requirements.txt
aenum==3.1.16
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
amqp==5.3.1
apscheduler==3.11.0
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==44.0.2
django==5.2
django-apscheduler==0.7.0
django-celery-results==2.6.0
django-debug-toolbar==4.3.0
django-environ==0.12.0
django-extensions==3.2.3
django-filter==25.1
django-picklefield==3.2
django-prometheus==2.3.1
django-q2==1.6.2
django-ratelimit==4.1.0
django-redis==5.4.0
djangorestframework==3.15.1
dnspython==2.7.0
drf-yasg==1.21.10
faker==37.1.0
flatdict==4.0.1
frozenlist==1.6.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
jwcrypto==1.5.6
kombu==5.5.3
loguru==0.7.3
mongoengine==0.29.1
multidict==6.4.3
numpy==2.2.5
okta==2.9.12
packaging==25.0
pandas==2.2.3
pluggy==1.5.0
prometheus-client==0.21.1
prompt-toolkit==3.0.51
propcache==0.3.1
pycparser==2.22
pycryptodomex==3.22.0
pyopenssl==24.3.0
pydash==8.0.5
pyjwt==2.10.1
pymongo==4.12.1
pytest==8.3.5
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.2.1
pytz==2025.2
pyyaml==6.0.2
redis==5.2.1
requests==2.32.3
sentry-sdk==2.27.0
six==1.17.0
sqlparse==0.5.3
typing-extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
win32-setctime==1.2.0
xmltodict==0.14.2
yarl==1.20.0